[AlertService] General refactor and implemented alerts for all sensors.

+++ AlertService/Notifiers/INotifier.cs
@@ -4,6 +4,6 @@ namespace Com.AlertService.Notifiers
 {
     public interface INotifier
     {
-        void Notify(double value, string sensorType, AlertType alertType);
+        void Notify(float value, string sensorType, AlertType alertType);
     }
 }
